feat: update README and HTML metadata, add robots.txt and sitemap.xml, refactor imports for consistency
