+++ index.html
@@ -1,13 +1,36 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/assets/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Snapgram</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
+<meta charset="UTF-8" />
+<link rel="icon" type="image/svg+xml" href="/assets/favicon.ico" />
+<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+<title>Modern Snapgram - Share Your Moments Instantly</title>
+<meta name="description"
+  content="Modern Snapgram is a fast, modern platform to share and discover stories, moments, and photos with friends and the world." />
+<meta name="keywords" content="Snapgram, social media, stories, photos, sharing, instant, modern" />
+<meta name="robots" content="index,follow" />
+<meta property="og:title" content="Modern Snapgram - Share Your Moments Instantly" />
+<meta property="og:description"
+  content="Share and discover stories, moments, and photos instantly with Modern Snapgram." />
+<meta property="og:type" content="website" />
+<meta property="og:url" content="https://modern-snapgram.vercel.app/" />
+<meta property="og:image" content="https://modern-snapgram.vercel.app/assets/images/side-img.svg" />
+<meta name="twitter:card" content="summary_large_image" />
+<meta name="twitter:title" content="Modern Snapgram - Share Your Moments Instantly" />
+<meta name="twitter:description"
+  content="Share and discover stories, moments, and photos instantly with Modern Snapgram." />
+<meta name="twitter:image" content="https://modern-snapgram.vercel.app/assets/images/side-img.svg" />
+<script type="application/ld+json">
+    {
+      "@context": "https://schema.org",
+      "@type": "WebSite",
+      "name": "Modern Snapgram",
+      "url": "https://modern-snapgram.vercel.app/"
+    }
+  </script>
+
+<body>
+  <div id="root"></div>
+  <script type="module" src="/src/main.tsx"></script>
+</body>
+
 </html>
