+++ public/robots.txt
@@ -1,0 +1,5 @@
+User-agent: *
+Allow: /
+Allow: /sign-in
+Allow: /sign-up
+Sitemap: https://modern-snapgram.vercel.app/sitemap.xml
