+++ src/components/shared/posts/UserPosts.tsx
@@ -1,7 +1,7 @@
 import InfiniteScroll from '@/components/shared/app/InfiniteScroll'
 import GridPostList from '@/components/shared/posts/GridPostList'
 import GridPostSkeleton from '@/components/shared/skeletons/GridPostSkeleton'
-import { useGetInfiniteUserPosts } from '@/states/query/hooks/infiniteQueries'
+import { useGetInfiniteUserPosts } from '@/states/TanStack-query/hooks/infinite-queries/posts/useGetInfiniteUserPosts'
 import { useMemo } from 'react'
 
 const UserPosts = ({ userId }: { userId: string }) => {
