+++ src/components/shared/profile/Followers.tsx
@@ -1,7 +1,7 @@
 import InfiniteScroll from '@/components/shared/app/InfiniteScroll'
 import AllUsersSkeleton from '@/components/shared/skeletons/AllUsersSkeleton'
 import UserCard from '@/components/shared/users/UserCard'
-import { useGetInfiniteFollowers } from '@/states/query/hooks/infiniteQueries'
+import { useGetInfiniteFollowers } from '@/states/TanStack-query/hooks/infinite-queries/followers/useGetInfiniteFollowers'
 import { E_USERS } from '@/values'
 import { useMemo } from 'react'
