+++ src/layouts/_root/pages/People.tsx
@@ -6,7 +6,7 @@ import SearchUsersResults from '@/components/shared/users/SearchUsersResults'
 import UserCard from '@/components/shared/users/UserCard'
 import { Input } from '@/components/ui/input'
 import { useDebounce } from '@/hooks/useDebounce'
-import { useGetInfiniteUsers } from '@/states/query/hooks/infiniteQueries'
+import { useGetInfiniteUsers } from '@/states/TanStack-query/hooks/infinite-queries/users/useGetInfiniteUsers'
 import { E_USERS } from '@/values'
 import { UserSearchIcon } from 'lucide-react'
 import { useMemo, useState } from 'react'
